Rethrow if RetryDecision.tryNextHost but there are no more hosts.
JAVA-709 (#372) added the capability for a RetryDecision to specify that
a request may be retried on different host by using
RetryDecision.tryNextHost.  In the case where tryNextHost is used and
there are no more available hosts a NoHostAvailableException (no hosts
were tried) is surfaced to the user.   This change instead surfaces the
response that initiated the RetryDecision instead of
NoHostAvailableException.

Added #should_rethrow_on_first_unavailable_if_there_are_no_more_hosts()
to demonstrate this change.